pyproject.toml
[project]
name = "podcast-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "atsphinx-audioplayer>=0.2.0",
    "atsphinx-feed>=0.1.1",
    "atsphinx-og-article>=0.2.1",
    "furo>=2024.8.6",
    "rst-package-refs>=0.1.0",
    "sphinx>=8.1.3",
]

[tool.uv]
dev-dependencies = [
    "esbonio>=0.16.5",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
